Phase 3.2: Extract CrawlState class to crawl_state.py
- Create doc_search/crawl_state.py module with CrawlState class
- CrawlState provides thread-safe crawl state management:
  - Visited URLs tracking
  - Pending URL queue
  - Failed URL retry counts
  - Crawl statistics
  - State persistence to disk

- Update crawler.py to import CrawlState from crawl_state
- Remove duplicate class definition from crawler.py
- Remove unused imports (deque no longer needed in crawler.py)

All 427 tests pass.
